+++ src/main/java/com/mllfjn/simyys/interactive/Interactive.java
@@ -82,7 +82,7 @@ public class Interactive {
     }
 
     // 普攻的对群体伤害
-    public AttackInfo[] attackTypical(Skill skill, List<Character> targets, int multiplier, AttackType attackType) {
+    public AttackInfo[] attackTypical(Skill skill, List<Character> targets, double multiplier, AttackType attackType) {
         AttackInfo[] attackInfos = new AttackInfo[targets.size()];
         for (int i = 0; i < targets.size(); i++) {
             AttackInfo attackInfo = AttackInfo.createTypicalAttack(owner, skill, targets.get(i), multiplier, attackType);
@@ -100,7 +100,7 @@ public class Interactive {
     }
 
     // 普通的对单体伤害
-    public AttackInfo attackTypical(Skill skill, Character target, int multiplier, AttackType attackType) {
+    public AttackInfo attackTypical(Skill skill, Character target, double multiplier, AttackType attackType) {
         AttackInfo attackInfo = AttackInfo.createTypicalAttack(owner, skill, target, multiplier, attackType);
         attack(attackInfo);
         return attackInfo;
